# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mural Pay Marketplace API",
    "version": "1.0.0",
    "description": "Backend API for a USDC-payment marketplace with automatic COP conversion via Mural Pay."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is running"
          }
        }
      }
    },
    "/merchant/orders": {
      "get": {
        "summary": "List all orders with payment status (merchant view)",
        "responses": {
          "200": {
            "description": "Array of orders with items and withdrawal status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "customerName": {
                        "type": "string"
                      },
                      "customerEmail": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "PENDING_PAYMENT",
                          "PAID",
                          "COMPLETED",
                          "FAILED"
                        ]
                      },
                      "totalUsdc": {
                        "type": "number"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/merchant/withdrawals": {
      "get": {
        "summary": "List all COP withdrawals (merchant view)",
        "responses": {
          "200": {
            "description": "Array of COP withdrawal records with payout status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "orderId": {
                        "type": "string"
                      },
                      "muralPayoutRequestId": {
                        "type": "string"
                      },
                      "amountCop": {
                        "type": "number",
                        "nullable": true,
                        "description": "COP amount received (null until payout is executed)"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "PENDING",
                          "PROCESSING",
                          "COMPLETED",
                          "FAILED"
                        ]
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Create a new order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "customerName",
                  "customerEmail",
                  "items"
                ],
                "properties": {
                  "customerName": {
                    "type": "string"
                  },
                  "customerEmail": {
                    "type": "string"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "productId",
                        "quantity"
                      ],
                      "properties": {
                        "productId": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created — use walletAddress, chainId, and tokenAddress to construct the USDC payment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string",
                      "description": "Use this to poll GET /orders/{id} for payment status"
                    },
                    "totalUsdc": {
                      "type": "number",
                      "description": "Exact USDC amount the customer must send"
                    },
                    "walletAddress": {
                      "type": "string",
                      "description": "Merchant Polygon wallet address (payment destination)"
                    },
                    "chainId": {
                      "type": "integer",
                      "description": "Polygon Amoy chain ID (80002)"
                    },
                    "tokenAddress": {
                      "type": "string",
                      "description": "USDC contract address on Polygon Amoy"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/orders/{id}/confirm": {
      "patch": {
        "summary": "Store the blockchain tx hash after wallet submission",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "txHash"
                ],
                "properties": {
                  "txHash": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tx hash stored"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get order status",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details and current payment status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "customerName": {
                      "type": "string"
                    },
                    "customerEmail": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING_PAYMENT",
                        "PAID",
                        "COMPLETED",
                        "FAILED"
                      ]
                    },
                    "totalUsdc": {
                      "type": "number"
                    },
                    "walletAddress": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "List all products",
        "responses": {
          "200": {
            "description": "Array of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "priceUsd": {
                        "type": "number"
                      },
                      "imageUrl": {
                        "type": "string"
                      },
                      "stock": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/mural": {
      "post": {
        "summary": "Receive Mural Pay webhook events",
        "responses": {
          "200": {
            "description": "Event received"
          },
          "401": {
            "description": "Invalid signature"
          }
        }
      }
    }
  },
  "tags": []
}